requirements-dev.txt
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
decorators==2.0.7
Django==5.2.13
django-celery-beat==2.9.0
django-celery-results==2.6.0
django-modeltranslation==0.20.3
django-environ==0.12.0
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
phonenumbers==9.0.28
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.12.1
python-crontab==3.3.0
pyTelegramBotAPI==4.25.0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
